@@ -1,17 +1,19 @@
 module github.com/omniscale/imposm3
 
-go 1.12
+go 1.21.0
 
 require (
-	github.com/fsnotify/fsnotify v1.4.7 // indirect
 	github.com/gogo/protobuf v1.3.2
 	github.com/jmhodges/levigo v1.0.1-0.20191214093932-ed89ec741d96
-	github.com/kr/pretty v0.1.0 // indirect
 	github.com/lib/pq v1.8.0
-	github.com/omniscale/go-osm v0.2.1
-	github.com/pkg/errors v0.8.0
-	golang.org/x/sys v0.0.0-20220818161305-2296e01440c6 // indirect
+	github.com/omniscale/go-osm v0.3.1
+	github.com/pkg/errors v0.9.1
+	gopkg.in/yaml.v2 v2.4.0
+)
+
+require (
+	github.com/kr/pretty v0.3.1 // indirect
+	golang.org/x/sys v0.4.0 // indirect
 	gopkg.in/check.v1 v1.0.0-20190902080502-41f04d3bba15 // indirect
-	gopkg.in/fsnotify.v1 v1.4.2 // indirect
-	gopkg.in/yaml.v2 v2.2.8
+	gopkg.in/fsnotify.v1 v1.4.7 // indirect
 )
