@@ -32,11 +32,20 @@ func estimateFromPBF(filename string, before time.Duration, replicationURL strin
 		}
 		timestamp = fstat.ModTime()
 	}
+	return estimateFromTimestamp(timestamp, before, replicationURL, replicationInterval)
+}
 
+func estimateFromTimestamp(timestamp time.Time, before time.Duration, replicationURL string, replicationInterval time.Duration) (*state.DiffState, error) {
 	if replicationURL == "" {
-		replicationURL = "https://planet.openstreetmap.org/replication/minute/"
+		switch replicationInterval {
+		case time.Hour:
+			replicationURL = "https://planet.openstreetmap.org/replication/hour/"
+		case time.Hour * 24:
+			replicationURL = "https://planet.openstreetmap.org/replication/day/"
+		default:
+			replicationURL = "https://planet.openstreetmap.org/replication/minute/"
+		}
 	}
-
 	seq, err := estimateSequence(replicationURL, replicationInterval, timestamp)
 	if err != nil {
 		return nil, errors.Wrap(err, "fetching current sequence for estimated import sequence")
