@@ -126,37 +126,23 @@ func TestExpireTiles(t *testing.T) {
 				{8239, 8009, 14},
 			}, true},
 
-			{"create polygon (outline)", []tile{
-				{8310, 8005, 14}, {8302, 7991, 14}, {8283, 7993, 14},
-				{8300, 8009, 14}, {8283, 8003, 14}, {8308, 8009, 14},
-				{8310, 7995, 14}, {8285, 8009, 14}, {8288, 8009, 14},
-				{8301, 8009, 14}, {8310, 8002, 14}, {8302, 8009, 14},
-				{8310, 8003, 14}, {8286, 8009, 14}, {8300, 7991, 14},
-				{8283, 7994, 14}, {8296, 8009, 14}, {8298, 8009, 14},
-				{8310, 8009, 14}, {8283, 7999, 14}, {8283, 7992, 14},
-				{8290, 7991, 14}, {8305, 8009, 14}, {8309, 7991, 14},
-				{8306, 7991, 14}, {8291, 7991, 14}, {8283, 7996, 14},
-				{8310, 7996, 14}, {8293, 7991, 14}, {8310, 8007, 14},
-				{8310, 8001, 14}, {8307, 8009, 14}, {8299, 8009, 14},
-				{8310, 7998, 14}, {8310, 7999, 14}, {8301, 7991, 14},
-				{8283, 7998, 14}, {8283, 8006, 14}, {8289, 8009, 14},
-				{8310, 8008, 14}, {8285, 7991, 14}, {8283, 8002, 14},
-				{8289, 7991, 14}, {8286, 7991, 14}, {8288, 7991, 14},
-				{8283, 8008, 14}, {8283, 8005, 14}, {8310, 7992, 14},
-				{8310, 8004, 14}, {8310, 7991, 14}, {8296, 7991, 14},
-				{8292, 7991, 14}, {8283, 8009, 14}, {8291, 8009, 14},
-				{8293, 8009, 14}, {8284, 8009, 14}, {8287, 7991, 14},
-				{8297, 8009, 14}, {8283, 8007, 14}, {8299, 7991, 14},
-				{8310, 7997, 14}, {8303, 8009, 14}, {8290, 8009, 14},
-				{8306, 8009, 14}, {8283, 7995, 14}, {8283, 8000, 14},
-				{8295, 8009, 14}, {8310, 8006, 14}, {8304, 8009, 14},
-				{8295, 7991, 14}, {8292, 8009, 14}, {8309, 8009, 14},
-				{8283, 8004, 14}, {8307, 7991, 14}, {8305, 7991, 14},
-				{8283, 8001, 14}, {8284, 7991, 14}, {8297, 7991, 14},
-				{8310, 7993, 14}, {8303, 7991, 14}, {8294, 8009, 14},
-				{8287, 8009, 14}, {8283, 7991, 14}, {8283, 7997, 14},
-				{8308, 7991, 14}, {8304, 7991, 14}, {8298, 7991, 14},
-				{8310, 8000, 14}, {8310, 7994, 14}, {8294, 7991, 14},
+			{"create polygon (zoom out)", []tile{
+				{2073, 2002, 12}, {2076, 1999, 12}, {2076, 2000, 12},
+				{2073, 2001, 12}, {2070, 2000, 12}, {2070, 2002, 12},
+				{2071, 2002, 12}, {2077, 1999, 12}, {2073, 1997, 12},
+				{2074, 1999, 12}, {2075, 2001, 12}, {2077, 1997, 12},
+				{2071, 1997, 12}, {2075, 1998, 12}, {2073, 2000, 12},
+				{2076, 1998, 12}, {2074, 1998, 12}, {2071, 2000, 12},
+				{2076, 1997, 12}, {2076, 2001, 12}, {2075, 2002, 12},
+				{2072, 1997, 12}, {2076, 2002, 12}, {2070, 1998, 12},
+				{2074, 2000, 12}, {2077, 2001, 12}, {2075, 1997, 12},
+				{2074, 1997, 12}, {2071, 2001, 12}, {2075, 1999, 12},
+				{2072, 1999, 12}, {2072, 2001, 12}, {2077, 2000, 12},
+				{2073, 1999, 12}, {2077, 2002, 12}, {2072, 2000, 12},
+				{2071, 1999, 12}, {2072, 1998, 12}, {2075, 2000, 12},
+				{2072, 2002, 12}, {2071, 1998, 12}, {2073, 1998, 12},
+				{2070, 1999, 12}, {2077, 1998, 12}, {2074, 2002, 12},
+				{2074, 2001, 12}, {2070, 1997, 12}, {2070, 2001, 12},
 			}, true},
 		} {
 
